% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maintenance_metabolism.R
\docType{data}
\name{Maintenance_metabolism}
\alias{Maintenance_metabolism}
\title{Example of Maintenance metabolism output}
\format{A data frame with 192 rows and 17 variables}
\usage{
data(Maintenance_metabolism)
}
\description{
Example of Maintenance metabolism output
}
\keyword{datasets}
