% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maintenance_no_tides.R
\docType{data}
\name{Maintenance_no_tides}
\alias{Maintenance_no_tides}
\title{Example of Maintenence metabolism output without considering tides}
\format{A data frame with 192 rows and 16 variables}
\usage{
data(Maintenance_no_tides)
}
\description{
Example of Maintenence metabolism output without considering tides
}
\keyword{datasets}
