% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmaint.R
\name{mmaint}
\alias{mmaint}
\title{Calculates maintenance metabolism}
\usage{
mmaint(weatherData, tide_table, timeBudget, breast_height, body_mass)
}
\arguments{
\item{weatherData}{a data frame with wind speed (named "wind_speed") in m/s,
temperature (named "temperature") in Celsius degrees, global horizontal
irradiance (named "GHI") in Wh/m^2, and datetime (named "date") in POSIXct
format.}

\item{tide_table}{a tide table based on which the tide codes will be calculated;
these codes will be used to merge with timeBudget; must have a column named
"date" with datetime info in POSIXct format and another one named "type"
with values 'High' or 'Low'.}

\item{timeBudget}{a data frame with tidal codes (column named "tide_code")
and the proportion of time spent on each microhabitat in columns 'h1', 'h2',
'h3' ... 'h11', corresponding to the microhabitats in Wiersma & Piersma
(1994; table 1). See correspondence under Details. If there is no variation
in relation to tides, provide a data frame with only one row and the
proportion per microhabitat}

\item{breast_height}{breast height, in meters.}

\item{body_mass}{body mass, in grams.}
}
\value{
a data frame with the date, weather data, tide codes (if used), time
  budget and a column with the maintenance metabolism in W; prints a plot
  with the initial 15 days of data and the range of m_maint in the data
}
\description{
Calculates the maintenance metabolism, in watt, for a shorebird following the
model of Wiersma & Piersma (1994); the thermal conductance used is calculated
following Kendeigth et al. (1977). Thermoregulatory costs will only occur when
maintenance metabolism is above the Basal Metabolic Rate.
}
\details{
Wiersma, P. and Piersma, T. 1994. Effects of microhabitat, flocking, climate
and migratory goal on energy expenditure in the annual cycle of red knots. -
Condor 96: 257 - 279.\cr Kendeigth, S. C., V. R. Dol'nik, and V. M. Gavrilov.
1977. Avian Energetics. Pp. 127 - 204 in J. Pinowski and S. C. Kendeigth,
editors. Granivorous birds in ecosystems. Cambridge University Press,
Cambridge, UK.

Microhabitats:\cr h1 - Densely vegetated salt marsh\cr h2 - Vegetated salt
marsh\cr h3 - Mudflat and bare salt marsh\cr h4 - On snow on tundra\cr h5 -
Behind tundra hillock\cr h6 - On tundra hillock\cr h7 - Dense group\cr h8 -
Open group\cr h9 - Solitary\cr h10 - Head in wind\cr h11 - Flank in wind\cr
}
\examples{
mmaint(weather, tides, time_budget, 0.17, 420)
}
