% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_data.R
\docType{data}
\name{summary_data}
\alias{summary_data}
\title{Example of summary data after using 'dplyr'}
\format{A data frame with 13 rows and 4 variables}
\usage{
data(summary_data)
}
\description{
Example of summary data after usung 'dplyr'
}
\keyword{datasets}
