% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_codes.R
\name{tidal_codes}
\alias{tidal_codes}
\title{Creates tidal codes}
\usage{
tidal_codes(observations, tide_table, round_digits = 1)
}
\arguments{
\item{observations}{data to which tide code will be appended; must have a
column named "date" with datetime info in POSIXct format.}

\item{tide_table}{a tide table upon tide codes will be calculated; must
have a column named "date" with datetime info in POSIXct format and
another column named "type" with values 'High' or 'Low'.}

\item{round_digits}{defines the accuracy of tide codes. If round_digits = 1
(default), observations are coded per hour. If round_digits = 0.5, creates
half-hour tide codes.}
}
\value{
The \code{observations} data frame entered with a new column named
  'tide_code'.
}
\description{
Creates a numerical code for the tide state in relation to the closest low
tide peak.
}
\details{
For each observation, the time difference to the nearest low tide is
calculated (in hours) and coded into a numerical category where 0
represents low tide and 6 high tide. If tide is ebbing, values are negative.
If tide is rising values are positive.

Note: To ensure tide code is calculated accurately for the observations given,
you must ensure that observation dates contain a time zone attribute. You can
define this with the field \code{tz} in \code{\link{as.POSIXct}}. For help with
accepted time zone codes, see \code{\link{OlsonNames}}.
}
\examples{
tidalCodes(observations, tides_january)
}
