% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides_jan.R
\docType{data}
\name{tides_jan}
\alias{tides_jan}
\title{Tide extremes at Iwik - Mauritania.}
\format{A data frame with 27 rows and 5 variables}
\usage{
data(tides_jan)
}
\description{
A dataset containing information on the date, time, height and type of tide extremes
 at Iwik, Mauritania between the 7th and 14th of January 2017. The variables are as follows:
}
\details{
\itemize{
  \item responseLon. longitude coordinates (decimal degrees) where tide data were collected
  \item responseLat. latitude coordinates (decimal degrees) where tide data were collected
  \item datetime. date and time (UTC) of tide extremes
  \item height. height of tide extremes (in meters)
  \item type. type of tide extreme (low or high)
}
}
\keyword{datasets}
