% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_budget.R
\docType{data}
\name{time_budget}
\alias{time_budget}
\title{Example of time budget recorded in the field}
\format{A data frame with 570 rows and 4 variables}
\usage{
data(time_budget)
}
\description{
A dataset containing observations of the proportion of roosting and not roosting individuals
}
\details{
\itemize{
  \item date. date when the observation was made
  \item time. time when the observation was made latitude coordinates (decimal degrees) where tide data were collected
  \item perc_roosting. proportion of individuals roosting
  \item perc_not_roosting. proportion of individuals not roosting
}
}
\keyword{datasets}
