% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_budget_no_tides.R
\docType{data}
\name{time_budget_no_tides}
\alias{time_budget_no_tides}
\title{Example of proportion of each microhabitat use, irrespective of tide}
\format{A data frame with 1 rows and 2 variables}
\usage{
data(time_budget_no_tides)
}
\description{
A dataset containing hypothetical proportions of microhabitat use,
irrespective of tide
}
\details{
\itemize{ \item h3 and h8. codes to the microhabitats mentioned in Wiersma &
Piersma (1994). See \code{\link{mmaint}} details.
}
}
\keyword{datasets}
