% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_budget_with_code.R
\docType{data}
\name{time_budget_with_code}
\alias{time_budget_with_code}
\title{Example of time budget data with a code for tide state}
\format{A data frame with 590 rows and 4 variables}
\usage{
data(time_budget_with_code)
}
\description{
Example of time budget data with a code for tide state
}
\keyword{datasets}
