% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_jan.R
\docType{data}
\name{weather_jan}
\alias{weather_jan}
\title{Weather data from Iwik - Mauritania.}
\format{A data frame with 192 rows and 4 variables}
\usage{
data(weather_jan)
}
\description{
A dataset containing hourly information on the temperature, wind speed and irradiation 
 at Iwik, Mauritania between the 7th and 14th of January 2017. The variables are as follows:
}
\details{
\itemize{
  \item date. date and time (UTC) of measurement
  \item temperature. Estimates at 183 cm (~6 ft) above ground, provided in degrees Celsius
  \item wind_speed. Estimates at 10m above ground, provided in m/s
  \item GHI. Estimates of horizontal irradiation at ground level, provided in in Wh/m2
}
}
\keyword{datasets}
